requirements.txt
RPi.GPIO
PyYAML